Add LatLng to tinylang
